Move components/shared/map-select.js -> components/shared/MapSelect.js
